+++ packages/react/src/components/DataTable/src/components/DataTableDropdownAction/index.ts
@@ -1,0 +1,2 @@
+export { DataTableDropdownAction } from "./DataTableDropdownAction";
+export type { DataTableDropdownActionProps } from "./dataTableDropdownAction.types";